fixed pytest fixtures placement, made create template api use template services instead of message_bus
